requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.6.3
